Harmless renames needed for the variable viewer
