+++ src/kernels/variables/jupyterVariables.ts
