+++ src/test/datascience/debugLocationTracker.unit.test.ts
@@ -4,7 +4,7 @@
 /* eslint-disable , , @typescript-eslint/no-explicit-any, no-multi-str, no-trailing-spaces */
 import { expect } from 'chai';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { DebugLocationTracker } from '../../platform/debugger/debugLocationTracker.node';
+import { DebugLocationTracker } from '../../platform/debugger/debugLocationTracker';
 import { IDebugLocation } from '../../platform/debugger/types';
 
 suite('Debug Location Tracker', () => {
