+++ src/webviews/extension-side/variablesView/variableViewActivationService.ts
